requirements.txt
torch==1.2.0
transformers==2.3.0
scipy==1.4.1
numpy==1.16.4
joblib==0.13.2
nltk==3.4.5
ConfigArgParse==1.0
sklearn_crfsuite==0.3.6
seqeval==0.0.5
spacy==2.2.3
Unidecode==1.1.1
tqdm==4.43.0
pandas==1.0.1
ipython==7.13.0
ptvsd==4.3.2
scikit_learn==0.22.2.post1
tensorboardX==2.0
